Completed Regression analysis
Updated analysis maps to include state vs county data.

Also included correlations between the different variables and found that not only does number of cases correlate with total population, but deaths due to covid correlates with number of vaccine doses and boosters.  This is a negative correlation with the percent vaccine coverage.